JCF: Issue #100: various changes with an eye toward better compliance with our coding standards, as well as quieting linter complaints about unavoidable unsigned integer use

+++ include/detdataformats/DAQEthHeader.hpp
@@ -26,9 +26,9 @@ struct DAQEthHeader
   word_t version : 6, det_id : 6, crate_id : 10, slot_id : 4, stream_id : 8, reserved : 6, seq_id : 12, block_length : 12;
   word_t timestamp { std::numeric_limits<word_t>::max() };
 
-  uint64_t get_timestamp() const // NOLINT(build/unsigned)
+  uint64_t get_timestamp() const // NOLINT(build/unsigned) maintain a consistent interface with DAQHeader
   {
-    return uint64_t(timestamp); // NOLINT(build/unsigned)
+    return timestamp;
   } 
 };
